sql - import from database

+++ src/nasuta-sebastian/Fsd.Sebastian.Cs/Fsd.Sebastian.Cs.Ex4.Web/Attributes/GreaterThanZeroAttribute.cs
@@ -7,7 +7,7 @@ namespace Fsd.Sebastian.Cs.Ex4.Web.Attributes
     {
         public override bool IsValid(object value)
         {
-            return Convert.ToInt32(value) > 0;
+            return Convert.ToDecimal(value) > 0;
         }
     }
 }
